fix: 参数校验

+++ src/main/java/com/axr/stockmanage/controller/ProductController.java
@@ -29,6 +29,9 @@ public class ProductController {
 
     @PostMapping("delete")
     public Result<Integer> deleteProduct(Integer id) {
+        if (id == null) {
+            throw new BusinessException("商品ID不能为空");
+        }
         return Result.success(productService.deleteProduct(id));
     }
